# contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/pay-periods:
    get:
      description: Requires payrun:view. Pre-loaded 2026 schedule.
      operationId: PayPeriodController_list_v1
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayPeriodResponse"
        "400":
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401":
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403":
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409":
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422":
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security:
        - bearer: []
      summary: List pay periods
      tags:
        - Pay Run & Holdback
  /v1/pay-runs:
    get:
      description: Requires payrun:view.
      operationId: PayRunController_list_v1
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayRunSummaryResponse"
        "400": &a123
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a124
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a125
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a126
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a127
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a128
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a129
        - bearer: []
      summary: List pay runs
      tags: &a130
        - Pay Run & Holdback
    post:
      description: Requires payrun:create. Computes preview lines via the engine;
        nothing frozen.
      operationId: PayRunController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayRunDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayRunResponse"
        "400": *a123
        "401": *a124
        "403": *a125
        "404": *a126
        "409": *a127
        "422": *a128
      security: *a129
      summary: Create / refresh a DRAFT pay run
      tags: *a130
  /v1/pay-runs/{id}:
    get:
      description: Requires payrun:view (scoped lines).
      operationId: PayRunController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayRunResponse"
        "400": *a123
        "401": *a124
        "403": *a125
        "404": *a126
        "409": *a127
        "422": *a128
      security: *a129
      summary: Get a pay run + lines
      tags: *a130
  /v1/pay-runs/{id}/lines:
    get:
      description: Requires payrun:view (scoped).
      operationId: PayRunController_lines_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayRunLineResponse"
        "400": *a123
        "401": *a124
        "403": *a125
        "404": *a126
        "409": *a127
        "422": *a128
      security: *a129
      summary: Per-rep computed lines
      tags: *a130
  /v1/pay-runs/{id}/lines/{lineId}/bonus:
    post:
      description: Requires payrun:approve. Draft only.
      operationId: PayRunController_setBonus_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: lineId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBonusDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayRunLineResponse"
        "400": *a123
        "401": *a124
        "403": *a125
        "404": *a126
        "409": *a127
        "422": *a128
      security: *a129
      summary: Set an ad-hoc bonus on a draft line
      tags: *a130
  /v1/pay-runs/{id}/finalize:
    post:
      description: "Requires payrun:approve. ATOMIC + IDEMPOTENT: freezes snapshots,
        pays sales, records/releases holdback (release timing PROPOSED — SRS
        §17), applies bonuses, composes net. Retry is a no-op."
      operationId: PayRunController_finalize_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          description: Safe to retry; finalize is idempotent.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayRunResponse"
        "400": *a123
        "401": *a124
        "403": *a125
        "404": *a126
        "409": *a127
        "422": *a128
      security: *a129
      summary: Finalize a pay run (the money action)
      tags: *a130
  /v1/pay-runs/{id}/export:
    post:
      description: Requires payrun:export. Configurable format; marks the run
        exported; recorded via audit.
      operationId: PayRunController_export_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportPayRunDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResultResponse"
        "400": *a123
        "401": *a124
        "403": *a125
        "404": *a126
        "409": *a127
        "422": *a128
      security: *a129
      summary: Generate the ADP export for a finalized run
      tags: *a130
  /v1/holdback-ledger:
    get:
      description: Requires payrun:view. Holds, schedule, release status.
      operationId: HoldbackLedgerController_list_v1
      parameters:
        - name: rep_id
          required: false
          in: query
          schema:
            type: string
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - held
              - scheduled
              - released
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HoldbackLedgerResponse"
        "400":
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401":
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403":
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409":
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422":
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security:
        - bearer: []
      summary: Holdback ledger
      tags:
        - Pay Run & Holdback
...
info:
  title: Redwave ERP / HRM API
  description: Versioned REST contract for the Redwave platform. Each protected
    endpoint notes the required (module, action) permission enforced server-side
    by the RBAC guard.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ErrorBodyDto:
      type: object
      properties:
        code:
          type: string
          example: UNPROCESSABLE_ENTITY
          description: Stable machine code for the error.
        message:
          type: string
          example: tier brackets must be contiguous (each min_count = previous max_count +
            1)
        details:
          type: object
          description: Optional structured detail (e.g. validation messages, an
            unpriced-products list).
          additionalProperties: true
      required:
        - code
        - message
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBodyDto"
      required:
        - error
...
    PayPeriodResponse:
      type: object
      properties:
        id:
          type: string
        period_number:
          type: number
          example: 3
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        payday:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - open
            - closed
            - paid
      required:
        - id
        - period_number
        - start_date
        - end_date
        - payday
        - status
    PayRunSummaryResponse:
      type: object
      properties:
        id:
          type: string
        pay_period_id:
          type: string
        pay_period:
          $ref: "#/components/schemas/PayPeriodResponse"
        run_date:
          type: string
          format: date-time
        status:
          type: string
          enum: &a204
            - draft
            - finalized
            - exported
        executed_by:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - pay_period_id
        - pay_period
        - run_date
        - status
        - executed_by
        - created_at
    CreatePayRunDto:
      type: object
      properties:
        pay_period_id:
          type: string
          description: The pay period to run (must be pre-loaded).
      required:
        - pay_period_id
    RepLiteResponse:
      type: object
      properties:
        id:
          type: string
        rep_code:
          type: string
        full_name:
          type: string
      required:
        - id
        - rep_code
        - full_name
    PayRunLineResponse:
      type: object
      properties:
        id:
          type: string
        pay_run_id:
          type: string
        rep_id:
          type: string
        rep:
          $ref: "#/components/schemas/RepLiteResponse"
        commission_70:
          type: string
          example: "2317.00"
          description: Decimal string. 70% advance (engine).
        holdback_release_30:
          type: string
          example: "993.00"
          description: Decimal string. 30% released from prior periods.
        expense_total:
          type: string
          example: "0.00"
          description: Decimal string. Approved expenses.
        incentive_total:
          type: string
          example: "0.00"
          description: Decimal string. Incentives (paid in full).
        bonus_amount:
          type: string
          example: "0.00"
          description: Decimal string. Ad-hoc bonus.
        bonus_note:
          type: string
          nullable: true
        clawback_total:
          type: string
          example: "0.00"
          description: Decimal string. Flat clawback deductions.
        net_payout:
          type: string
          example: "2317.00"
          description: Decimal string. Net = 70 + 30 + exp + inc + bonus − clawback (may
            be negative).
      required:
        - id
        - pay_run_id
        - rep_id
        - rep
        - commission_70
        - holdback_release_30
        - expense_total
        - incentive_total
        - bonus_amount
        - bonus_note
        - clawback_total
        - net_payout
    PayRunResponse:
      type: object
      properties:
        id:
          type: string
        pay_period_id:
          type: string
        pay_period:
          $ref: "#/components/schemas/PayPeriodResponse"
        run_date:
          type: string
          format: date-time
        status:
          type: string
          enum: *a204
        executed_by:
          type: string
        created_at:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: "#/components/schemas/PayRunLineResponse"
      required:
        - id
        - pay_period_id
        - pay_period
        - run_date
        - status
        - executed_by
        - created_at
        - lines
    SetBonusDto:
      type: object
      properties:
        amount:
          type: string
          example: "50.00"
          description: Bonus amount — exact decimal STRING, never a float.
        note:
          type: string
          description: Reason / note for the bonus.
      required:
        - amount
    ExportPayRunDto:
      type: object
      properties:
        format:
          type: string
          enum:
            - csv
            - json
          default: csv
          description: Export format.
    ExportResultResponse:
      type: object
      properties:
        pay_run_id:
          type: string
        format:
          type: string
          enum:
            - csv
            - json
          example: csv
        line_count:
          type: number
          example: 12
        content:
          type: string
          description: The rendered export payload (CSV text or JSON string).
      required:
        - pay_run_id
        - format
        - line_count
        - content
    HoldbackLedgerResponse:
      type: object
      properties:
        id:
          type: string
        rep_id:
          type: string
        origin_pay_period_id:
          type: string
        amount_held:
          type: string
          example: "993.00"
          description: Decimal string. The 30% held this origin period.
        scheduled_release_period_id:
          type: string
          nullable: true
        release_status:
          type: string
          enum:
            - held
            - scheduled
            - released
        released_in_pay_run_id:
          type: string
          nullable: true
        amount_released:
          type: string
          nullable: true
          example: "993.00"
          description: Decimal string. Amount released.
        clawback_applied:
          type: string
          nullable: true
          example: "0.00"
          description: Decimal string. Clawback applied against the hold.
      required:
        - id
        - rep_id
        - origin_pay_period_id
        - amount_held
        - scheduled_release_period_id
        - release_status
        - released_in_pay_run_id
        - amount_released
        - clawback_applied
...